+++ examples/data-sources/exoscale_instance_pool_list/data-source.tf
@@ -4,6 +4,6 @@ data "exoscale_instance_pool_list" "my_instance_pool_list" {
 
 output "my_instance_pool_ids" {
   value = join("\n", formatlist(
-    "%s", exoscale_instance_pool_list.my_instance_pool_list.pools.*.id
+    "%s", data.exoscale_instance_pool_list.my_instance_pool_list.pools.*.id
   ))
 }